+++ host2str.c
@@ -345,7 +345,7 @@ ldns_rdf2buffer_str_str(ldns_buffer *output, const ldns_rdf *rdf)
 	ldns_buffer_printf(output, "\"");
 	for (i = 1; i <= length; ++i) {
 		char ch = (char) data[i];
-		if (isprint(ch)) {
+		if (isprint((int)ch)) {
 			if (ch == '"' || ch == '\\') {
 				ldns_buffer_printf(output, "\\");
 			}
